Initial tool (#4)
## Summary

Implement the infrastructure of the tool.

This includes the diagnostics, our first rule `typing-any-used`.

This picks up on examples where a function definition has parameters
annotated with `Any`. Currently we only detect this if it is like the
following:

```py
def foo(x: Any): ...

# but not

def foo(x: Any | None): ...
```

This is because the `Any` is nested in the expr, to do this fully we
must walk the expression to check for `Any`.


The tool now works with `cargo run`.

Given a file like:

```py
from typing import Any

def foo(x: Any) -> None: ...
```

We get the stdout:

```shell
error[typing-any-used]: Using `typing.Any` in type annotations can lead to runtime errors.
 --> test.py:4:12
  |
4 | def foo(x: Any) -> None: ...
  |            ^^^
  |
info: rule `typing-any-used` is enabled by default

```


This is our first unsoundness rule.

Much of the diagnostics output was taken from the ruff codebase, they
have a very mature and easy api for emitting informative and useful
diagnostics.

We also use the ruff python ast parser here. This is incredibly fast and
it is very easy to use.

We also use `ty_python_semantic` crate from the ruff codebase too. This
allows us to infer the type of expressions and here we use it to infer
the type of the annotations which allows us to detect `Any` annotations.

## Test Plan

We have added a few snapshot tests to test the diagnostic output.

I expect to refactor this in the future but this is okay for just now