Add docs (#19)
## Summary

We make a few changes here. 

First we now have multiple crates here. So we can have a `dev` create
for generating documentation files.

And we add an auto generated `rules.md` file showing all the rules we
have now.

We also add mkdocs documentation. These docs are almost empty right now
but we will add more in the future.