Allow subtype assignments in globals mutation and typing cast (#58)
## Summary

Previously we emitted an error even when you used globals cast for an
assignment that was type safe. And the same for cast.

Here we fix that and don't emit diagnostics when the user makes a valid
cast or uses globals in a type safe way.

Resolves #57 

## Test Plan

Update snapshot test files.